% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{data_example_one}
\alias{data_example_one}
\title{A simulated dataset to showcase the usage of the MR_AHC function.}
\format{
A data frame with 90 rows and 4 variables:
\describe{
  \item{est_x}{the SNP-exposure associations}
  \item{se_x}{standard error of the SNP-exposure associations}
  \item{est_y}{the SNP-outcome associations}
  \item{se_y}{standard error of the SNP-outcome associations}
}
}
\usage{
data_example_one
}
\description{
A simulated dataset to showcase the usage of the MR_AHC function.
}
\keyword{datasets}
