% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{data_example_two}
\alias{data_example_two}
\title{A simulated dataset to showcase the usage of the MR_AHC_two function.}
\format{
A data frame with 90 rows and 6 variables:
\describe{
  \item{est_x}{the SNP-exposure associations}
  \item{se_x}{standard error of the SNP-exposure associations}
  \item{est_y1}{the SNP-outcome associations}
  \item{se_y1}{standard error of the SNP-outcome associations}
  \item{est_y2}{the SNP-outcome associations}
  \item{se_y2}{standard error of the SNP-outcome associations}
}
}
\usage{
data_example_two
}
\description{
A simulated dataset to showcase the usage of the MR_AHC_two function.
}
\keyword{datasets}
