% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_AHC.R
\name{MR_AHC}
\alias{MR_AHC}
\title{Implementing the MR-AHC method for detecting genetic variants clusters
and conducting post-clustering estimation using summary statistics in Mendelian randomization with one outcome.}
\usage{
MR_AHC(
  betaX,
  betaY,
  seX,
  seY,
  n,
  alpha = 0.05,
  tuning = 0.1/log(n),
  weak = TRUE,
  smallcluster = 4,
  outremove = FALSE,
  iter = FALSE,
  iter.p = 0.05
)
}
\arguments{
\item{betaX:}{A numeric vector of the SNP-exposure associations (p by 1 vector).}

\item{betaY:}{A numeric vector of the SNP-outcome associations (p by 1 vector).}

\item{seX:}{A numeric vector of the standard errors of the SNP-exposure associations (p by 1 vector).}

\item{seY:}{A numeric vector of the standard errors of the SNP-outcome associations (p by 1 vector).}

\item{n:}{A numeric scalar specifying the sample size.
In a two-sample MR design, we recommend using the sample size of the outcome sample.}

\item{alpha:}{A numeric scalar between 0 and 1 specifying the significance level
for the confidence interval (default = 0.05).}

\item{tuning:}{A numeric scalar specifying the threshold p-value for the Q test (default = 0.1/log(n)).}

\item{weak:}{Logical. If weak = TRUE, the weak IV robust Q test is used (default = TRUE).}

\item{smallcluster:}{A numeric scalar specifying the cutoff value for small clusters (default = 4).}

\item{outremove:}{Logical. If outremove = TRUE, conduct the outlier removal procedure (default = FALSE).}

\item{iter:}{Logical. When outremove = TRUE, if iter = TRUE, conduct the iterated
outlier removal procedure (default = FALSE).}

\item{iter.p:}{A numeric scalar specifying the threshold p-value for the individual Q in
the outlier removal procedure.}
}
\value{
Cluster_number: The number of clusters detected by the algorithm (including the junk cluster).

Cluster_number_real: The number of substantive clusters detected by the algorithm (excluding the junk cluster).

AHC_cluster: All the clusters detected by the algorithm. SNPs identities (as in the dataset) are listed in each cluster.

AHC_cluster_real: All the substantive clusters detected by the algorithm. SNPs identities (as in the dataset) are listed in each cluster.

Null_cluster: Cluster index of the null cluster.

Junk_cluster: Identities of SNPs in the junk cluster.

F: the F statistic for all the IVs.

AHC_results: A matrix that summarizes the clustering and estimation results, including:
          (a) length: The number of IVs in each cluster.
          (b) beta: the point estimate estimated with each cluster.
         (c) se: the standard error for the causal estimate in each cluster
         (d) t: the t statistic.
          (e) Qp: The p-value for the Q test of the instruments in each cluster.
          (f) I^2: The I statistic of the instruments in each cluster.

Confidence interval: The 95% (default) confidence intervals of the cluster-specific estimates.
}
\description{
Implementing the MR-AHC method for detecting genetic variants clusters
and conducting post-clustering estimation using summary statistics in Mendelian randomization with one outcome.
}
