% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_AHC_two.R
\name{MR_AHC_two}
\alias{MR_AHC_two}
\title{Implementing the MR-AHC method for detecting genetic variants clusters
and conducting post-clustering estimation using summary statistics in Mendelian randomization with two outcomes.}
\usage{
MR_AHC_two(
  betaX,
  betaY1,
  betaY2,
  seX,
  seY1,
  seY2,
  n,
  alpha = 0.05,
  tuning = 0.1/log(n),
  smallcluster = 4,
  outremove = FALSE,
  iter = FALSE,
  iter.p = 0.05,
  rho = 0
)
}
\arguments{
\item{betaX:}{A numeric vector of SNP-exposure associations (p by 1 vector).}

\item{betaY1:}{A numeric vector of SNP-outcome associations for outcome 1 (p by 1 vector).}

\item{betaY2:}{A numeric vector of SNP-outcome associations for outcome 2 (p by 1 vector).}

\item{seX:}{A numeric vector of the standard errors of the SNP-exposure associations (p by 1 vector).}

\item{seY1:}{A numeric vector of the standard errors of the SNP-outcome associations for outcome 1 (p by 1 vector).}

\item{seY2:}{A numeric vector of the standard errors of the SNP-outcome associations for outcome 2 (p by 1 vector).}

\item{n:}{A numeric scalar specifying the sample size. In a multi-sample MR design, we recommend using the smaller sample size of the outcome samples.}

\item{alpha:}{A numeric scalar between 0 and 1 specifying the significance level for the confidence interval (default = 0.05).}

\item{tuning:}{A numeric scalar specifying the threshold p-value for the Q test (default = 0.1/log(n)).}

\item{smallcluster:}{A numeric scalar specifying the cutoff value for small clusters (default = 4).}

\item{outremove:}{Logical. If outremove = TRUE, conduct the outlier removal procedure (default = FALSE).}

\item{iter:}{Logical. When outremove = TRUE, if iter = TRUE, conduct the iterated outlier removal procedure (default = FALSE).}

\item{iter.p:}{A numeric scalar specifying the threshold p-value for the individual Q in the outlier removal procedure.}

\item{rho:}{A numeric scalar specifying the covariance between the two SNP-specific Wald estimates for the two outcomes.
If the two outcome samples are non-overlapping, set rho = 0 (default = 0).}
}
\value{
Cluster_number: The number of clusters detected by the algorithm (including the junk cluster).

Cluster_number_real: The number of substantive clusters detected by the algorithm (excluding the junk cluster).

AHC_cluster: All the clusters detected by the algorithm. SNPs identities (as in the dataset) are listed in each cluster.

AHC_cluster_real: All the substantive clusters detected by the algorithm. SNPs identities (as in the dataset) are listed in each cluster.

Null_cluster_1: Cluster index of the null cluster for outcome 1.

Null_cluster_2: Cluster index of the null cluster for outcome 2.

Junk_cluster: Identities of SNPs in the junk cluster.

F: the F statistic for all the IVs.

AHC_results: A matrix that summarizes the clustering and estimation results, including:
          (a) length: The number of IVs in each cluster.
          (b) beta: the point estimates estimated with each cluster.
          (c) se: the standard errors for the causal estimates in each cluster
          (d) t: the t statistic.
          (e) Qp: The p-value for the Q test of the instruments in each cluster.
          (f) I^2: The I statistic of the instruments in each cluster.

Confidence interval: The 95% (default) confidence intervals of the cluster-specific estimates.
}
\description{
Implementing the MR-AHC method for detecting genetic variants clusters
and conducting post-clustering estimation using summary statistics in Mendelian randomization with two outcomes.
}
